% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages-json.R
\name{clone_gh_org_repos}
\alias{clone_gh_org_repos}
\title{Clone or update all repositories defined in 'packages.json'}
\usage{
clone_gh_org_repos(pkgs_json = NULL)
}
\arguments{
\item{pkgs_json}{Local path to 'packages.json' as created or updated by
running \link{om_packages_json}. That function must be run first, prior to
calling this function!}
}
\value{
Function primarily called for side-effect of clone or updating all
repositories defined in 'packages.json', but does invisibly return a vector
of paths to all local repositories of R packages as listed in \code{pkgs_json}.
}
\description{
Clone or update all repositories defined in 'packages.json'
}
