% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages-json.R
\name{om_packages_json}
\alias{om_packages_json}
\title{Collate and write the 'package.json' file for org}
\usage{
om_packages_json(org_path = NULL)
}
\arguments{
\item{org_path}{Path to root directory of organization repositories. Should
contain sub-directories for different GitHub organizations. These
sub-directories may be initially empty, and will be populated by the
(currently interanl) function, \code{clone_gh_org_repos()}.}
}
\value{
Path to 'packages.json' file containing data on all repositories
within organization(s).
}
\description{
Collate and write the 'package.json' file for org
}
