% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-dashboard-org.R
\name{orgmetrics_dashboard}
\alias{orgmetrics_dashboard}
\title{Start quarto dashboard with results of
\code{orgmetrics_collate_org_data} function for collation of data across orgs.}
\usage{
orgmetrics_dashboard(
  data_org,
  fn_calls,
  similarities,
  title = NULL,
  action = "preview"
)
}
\arguments{
\item{data_org}{Data on GitHub organization as returned from
\code{orgmetrics_collate_org_data} function.}

\item{fn_calls}{Data on function calls between packages of the specified
organization, as returned from the \code{rm_org_data_fn_call_network()} function.}

\item{similarities}{List of 'BM25' similarities between packages generated
by the 'pkgmatch' package}

\item{title}{If not \code{NULL} (default), a string specifying the organizational
title for the dashboard.}

\item{action}{One of "preview", to start and open a live preview of the
dashboard website, "render" to render a static version without previewing
or opening, or \code{NULL} to set up the quarto structure in the current
temporary directory without doing anything. This option is useful to
generate the dashboard structure so that it can be moved to a non-temporary
location, and deployed or previewed from there.}
}
\value{
(Invisibly) Path to main "index.html" document of quarto site. Note
that the site must be served with \code{action = "preview"}, and will not work by
simply opening this "index.html" file.
}
\description{
Start quarto dashboard with results of
\code{orgmetrics_collate_org_data} function for collation of data across orgs.
}
