% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{orgmetrics_deploy_r_univ}
\alias{orgmetrics_deploy_r_univ}
\title{Deploy a dashboard from an 'r-universe' "packages.json" file.}
\usage{
orgmetrics_deploy_r_univ(url, dest_dir = fs::path_temp())
}
\arguments{
\item{url}{URL of a '\if{html}{\out{<username>}}.r-universe.dev' GitHub repository,
containing these two files:
\item a "packages.json" file defining the repositories contained within the
'r-universe'
\item An additional "orgmetrics-config.yaml" file generated by
\link{orgmetrics_generate_yaml}.

\enumerate{
}}
}
\description{
Deploy a dashboard from an 'r-universe' "packages.json" file.
}
