% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{orgmetrics_generate_yaml}
\alias{orgmetrics_generate_yaml}
\title{Helper function to general a 'orgmetrics-config.yaml' file in an r-universe
repository otherwise containing only a "packages.json" file.}
\usage{
orgmetrics_generate_yaml(dest_dir, title = NULL, aggregation_period = 90)
}
\arguments{
\item{dest_dir}{Path to local clone of '\if{html}{\out{<username>}}.r-universe.dev' repository.}

\item{title}{Title for 'orgmetrics' dashboard.}

\item{aggregation_period}{Period in days over which prior activity is to be
aggregated.}
}
\description{
This function will create a new file in the location specified by
\code{dest_dir}, which must then be added to the repository in a commit, and
pushed to the remote location (such as GitHub).
}
