% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgmetrics-package.R
\docType{package}
\name{orgmetrics-package}
\alias{orgmetrics}
\alias{orgmetrics-package}
\title{orgmetrics: Metrics for Your GitHub Organization}
\description{
Metrics for your GitHub organization. Call one function to generate an interactive dashboard displaying the state of your organization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/orgmetrics/}
  \item \url{https://github.com/ropensci-review-tools/orgmetrics}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/orgmetrics/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
