% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{orgmetrics_deploy_r_univ}
\alias{orgmetrics_deploy_r_univ}
\title{Deploy a dashboard from an 'r-universe' "packages.json" file.}
\usage{
orgmetrics_deploy_r_univ(
  url = NULL,
  dest_dir = fs::path_temp(),
  title = NULL,
  aggregation_period = 365,
  action = NULL
)
}
\arguments{
\item{url}{URL of a '\if{html}{\out{<username>}}.r-universe.dev' GitHub repository,
containing a "packages.json" file defining the repositories contained within
the 'r-universe'.
If the parameter is not specified, function is presumed to be called within
local version of an r-universe-like repository.}

\item{dest_dir}{A local directory where all repositories of the specified
'r-universe' will be cloned, and also where data generated for the
'orgmetrics' dashboard will be stored. The default is a temporary path, but
a permanent local path is better for local usage of this function, to avoid
re-generating the same data each time.}

\item{title}{Title for 'orgmetrics' dashboard. Default is \code{NULL}, in
which case the title is taken to be the terminal element of \code{url}.}

\item{aggregation_period}{Period in days over which prior activity is to be
aggregated.}

\item{action}{One of "preview", to start and open a live preview of the
dashboard website, "render" to render a static version without previewing
or opening, or \code{NULL} to set up the quarto structure in the current
temporary directory without doing anything. This option is useful to
generate the dashboard structure so that it can be moved to a non-temporary
location, and deployed or previewed from there.}
}
\description{
Deploy a dashboard from an 'r-universe' "packages.json" file.
}
