% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{orgmetrics_deploy_r_univ}
\alias{orgmetrics_deploy_r_univ}
\title{Deploy a dashboard from an 'r-universe' "packages.json" file.}
\usage{
orgmetrics_deploy_r_univ(url, dest_dir = fs::path_temp())
}
\arguments{
\item{url}{URL of a '\if{html}{\out{<username>}}.r-universe.dev' GitHub repository,
containing these two files:
\enumerate{
\item a "packages.json" file defining the repositories contained within the
'r-universe'
\item An additional "orgmetrics-config.yaml" file generated by
\link{orgmetrics_generate_yaml}.
}}

\item{dest_dir}{A local directory where all repositories of the specified
'r-universe' will be cloned, and also where data generated for the
'orgmetrics' dashboard will be stored. The default is a temporary path, but
a permanent local path is better for local usage of this function, to avoid
re-generating the same data each time.}
}
\description{
Deploy a dashboard from an 'r-universe' "packages.json" file.
}
